requirements.txt
numpy
pandas
nmrglue
flask
matplotlib
scipy
pystan
